Extract search method into search endpoint
